refactor: move & unify several shared modules to make their roles more clear (#193)
* refactor: move issue-uuid to route-handler-call

* refactor: purge unused lib modules

* refactor: move remaining lib modules to util, since their roles are identical to util

* refactor: move unordered-list to ui

* refactor: move tag to ui

* refactor: purge unused preparing notice

* refactor: move skeleton to ui
