--- src/shared/lib/error.ts
@@ -1,5 +1,0 @@
-export function checkIsHaveErrorCode(
-  error: unknown
-): error is { code: string } {
-  return Boolean((error as { code: string }).code);
-}
