--- src/shared/notice/preparing.tsx
@@ -1,19 +1,0 @@
-export default function PreparingNotice() {
-  return (
-    <main className="w-full h-full max-w-3xl flex flex-col items-center px-4 sm:px-8 py-32 gap-8">
-      {/**
-       * @todo
-       * Add preparing notice image
-       */}
-      {/* <div className="w-full h-40 relative">
-        <Image
-          src={PreparingPng}
-          fill
-          alt="준비 중"
-          className="object-contain"
-        />
-      </div> */}
-      <h1 className="text-4xl font-bold">현재 준비 중입니다.</h1>
-    </main>
-  );
-}
